requirements.txt
boto3==1.36.2
dropbox==12.0.2
duckdb==1.1.3
orjson==3.10.15
pandas==2.2.3
python-dotenv==1.0.1
tqdm==4.67.1
